requirements.txt
Flask
python-dotenv
supabase
google-generativeai
pillow
gunicorn
